Fixes, renaming
